Fixed bridge assignment for cloning and also cleaned up some proxmox/cloning functions
